Allow clickable workspace names without munging them up front
As long as the key name is the first character of the workspace name, we
can do the necessary wrapping in the prettyprinter.  This lets will
eventually let us handle multiple bars cleanly, show workspace names
when switching, and so on.

It will also make it easy to redefine the clickwrap function as id on
old systems that don't support it.