shim for passing lj-update buffers to charm
It would be possible to make charm able to post directly from a file
without editing it; this would still be useful for mapping Access (which
lj-update uses) to Permission, and removing lj-update's separator line
("--text follows this line--").

In the mean time, it was a quick hack, and it works.