load vc instead of vc-dispatcher
vc-dispatcher is required by vc, so we get it anyway.  The reason for
preloading parts of the VC system is so that we can edit commit and
merge messages from the command line.

We want to edit messages in vc-log-edit-mode so that we can use the hook
to set things up differently from plain text-mode, in particular the
line length.