Make .xmonad and .xmonad/lib real directories
Machinery in the Makefile to recursively copy directories that start
with =, and symlink their contents.  That would be better as a script
that could be reused elsewhere, but it will work for now.  There are
some other places where it would be useful, too.

The .originals directory is moved to ~.  Ideally there shouldn't be any
system-dependent state or build products in ~/vv, so maybe the local and
opt subdirs should go there too.