Add special handling for .gtkrc-2.0 in install
.gtkrc-2.0 is treated differently between gnome (or at least gnome 3)
and xfce, and probably others.  In any case, it's subject to
modification by style editors.  We adapt the solution found in Mint/MATE
and put our own additions in an include file, .gtkrc-2.0.mine.  We then
append the include statement if it isn't there.  This will cause trouble
if something else has already written into the file; if that becomes a
problem we can use sed to prepend the include.