Comment out font properties in .Xdefaults.
They don't appear to be used anywhere except xterm; the original
references were to adobe-courier.  It's possible that one could replace
them with bitstream equivalents, but that would require figuring out
which package they're in.

It will eventually be worthwhile going through .Xdefaults and seeing
whether _anything_ is used anymore.