Add ws-terminal to run a "good" terminal
`ws-terminal` runs a terminal on a workspace.  It searches a list of
plausible candidate terminal emulators (searched in the order: command
line, environment variables, built-in defaults) and uses a set of
heuristics to figure out how to give the selected terminal a name.

By default, the name is ws$N-terminal, which tells xmonad to put it on
workspace $N.

By putting this logic in `ws-terminal`, we eliminate the need to have an
OS-specific value for goodTerminal in Local.hs.