Add bin/sshot, which takes a screenshot
If not given a filename on the command line it defaults to using
yyyy-mm-ddThh-mm-ss.png.  Eventually it should be able to supply the
extension if it's omitted.  It would also be nice if it took flags like
--help and -f (which tells xwd to include the WM frame).  When invoked
from, e.g., xmonad's mod-p menu, it is running in $HOME and can be given
paths accordingly.  Wouldn't hurt to default to ~/Pictures, either.