Copy .gitignore from dotfiles
If a .gitignore file is a symlink, git will warn about it but won't
follow the link.  Formerly, the top-level .gitignore file in Honu was a
symlink to `dotfiles/_gitignore_global`.  We don't actually need
everything in `_gitignore_global`, but having it doesn't hurt.  (NOTE:
This may need to be revisited at some point.)